fix: add scrolling to category articles list and refactor
Closes #474
