+++ mobile/src/components/categories/index.ts
@@ -1,0 +1,3 @@
+export * from "./CategoryFormComponent";
+export * from "./CategoryDetailsComponent";
+export * from "./CategoryArticlesComponent";
