+++ mobile/src/components/index.ts
@@ -16,5 +16,4 @@ export * from "./FadeInItem";
 export * from "./KeakitCRUDButton";
 export * from "./KeakitSearchBar";
 export * from "./KeakitTag";
-export * from "./CategoryFormComponent";
-export * from "./CategoryDetailsComponent";
+export * from "./categories";
